+++ src/hooks/dismissBrowser.ts
@@ -3,6 +3,8 @@ import * as WebBrowser from 'expo-web-browser';
 /**
  * Dismisses the web browser.
  */
-export const dismissBrowser = (): void => {
-  WebBrowser.dismissBrowser();
+export const dismissBrowser = async (): Promise<void> => {
+  setTimeout(async () => {
+    await WebBrowser.dismissBrowser();
+  }, 500);
 };