+++ src/hooks/login.ts
@@ -1,11 +1,13 @@
 import { toHex } from '@dfinity/agent';
-import { Ed25519KeyIdentityValueStorageWrapper } from '../storage/Ed25519KeyIdentityValueStorageWrapper';
 import { RedirectPathStorage } from '../storage/RedirectPathStorage';
 import { buildIIIntegrationURL } from './buildIIIntegrationURL';
 import { openBrowser } from './openBrowser';
 import { LoginOuterParams } from '../types';
 import { saveRedirectPath } from './saveRedirectPath';
 import { Ed25519KeyIdentity } from '@dfinity/identity';
+import { CryptoModule } from 'expo-crypto-universal';
+import { SessionIdStorage } from '../storage/SessionIdStorage';
+import { AppKeyStorage } from '../storage/AppKeyStorage';
 /**
  * Represents the parameters required for the login function.
  * @property {string} localIPAddress - The local IP address.
@@ -15,10 +17,12 @@ import { Ed25519KeyIdentity } from '@dfinity/identity';
  * @property {string} frontendCanisterId - The frontend canister ID.
  * @property {string} iiIntegrationCanisterId - The II integration canister ID.
  * @property {string} authPath - The authentication path.
- * @property {Ed25519KeyIdentityValueStorageWrapper} appKeyStorage - The storage wrapper for app key.
+ * @property {AppKeyStorage} appKeyStorage - The storage wrapper for app key.
  * @property {RedirectPathStorage} redirectPathStorage - The storage wrapper for redirect path.
+ * @property {SessionIdStorage} sessionIdStorage - The storage wrapper for session ID.
  * @property {string} currentPath - The current path.
  * @property {LoginOuterParams} loginOuterParams - The outer parameters for the login function.
+ * @property {CryptoModule} cryptoModule - The crypto module.
  */
 type LoginParams = {
   localIPAddress: string;
@@ -27,10 +31,12 @@ type LoginParams = {
   deepLink: string;
   frontendCanisterId: string;
   iiIntegrationCanisterId: string;
-  appKeyStorage: Ed25519KeyIdentityValueStorageWrapper;
+  appKeyStorage: AppKeyStorage;
   redirectPathStorage: RedirectPathStorage;
+  sessionIdStorage: SessionIdStorage;
   currentPath: string;
   loginOuterParams: LoginOuterParams;
+  cryptoModule: CryptoModule;
 };
 
 /**
@@ -51,8 +57,10 @@ export const login = async ({
   iiIntegrationCanisterId,
   appKeyStorage,
   redirectPathStorage,
+  sessionIdStorage,
   currentPath,
   loginOuterParams,
+  cryptoModule,
 }: LoginParams): Promise<void> => {
   try {
     console.log('Logging in');
@@ -66,6 +74,8 @@ export const login = async ({
     const appKey = await Ed25519KeyIdentity.generate();
     await appKeyStorage.save(appKey);
     const pubkey = toHex(appKey.getPublicKey().toDer());
+    const sessionId = toHex((await cryptoModule.getRandomBytes(32)).buffer);
+    await sessionIdStorage.save(sessionId);
 
     const iiIntegrationURL = buildIIIntegrationURL({
       pubkey,
@@ -75,6 +85,7 @@ export const login = async ({
       deepLink,
       frontendCanisterId,
       iiIntegrationCanisterId,
+      sessionId,
     });
 
     await openBrowser(iiIntegrationURL);